+++ main/proxyhttp/config.go
@@ -1,0 +1,32 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package proxyhttp
+
+import (
+	"context"
+	"net/http"
+
+	filter "github.com/DomZippilli/gcs-proxy-cloud-function/main/filter"
+)
+
+// This function will be called in main.go for GET requests
+func GET(ctx context.Context, output http.ResponseWriter, input *http.Request) {
+	get(ctx, output, input, []filter.MediaFilter{})
+}
+
+// func HEAD
+
+// func POST
+
+// func DELETE
